playing w some architecture designs, but prototype redux layer should be working
