--- src/gameclient/ui/context/actions.ts
@@ -1,15 +1,0 @@
-export interface Actions {
-    [key: string]: {
-        type: ActionTypes
-    }
-}
-
-export enum ActionTypes {
-    INCREMENT = 'INCREMENT',
-    DECREMENT = 'DECREMENT'
-}
-
-export const actions = {
-    increment: { type: ActionTypes.INCREMENT },
-    decrement: { type: ActionTypes.DECREMENT }
-};
