+++ src/gameclient/ui/store/actions.ts
@@ -1,0 +1,12 @@
+import { ChatMessageData } from "../../../packets/data/chatmessagedata";
+
+export type ChatHistoryState = {
+    chatHistory: ChatMessageData[]
+}
+
+export type ChatHistoryAction = {
+    type: string
+    chatMessageData: ChatMessageData
+}
+
+// type DispatchType = (args: ChatHistoryAction) => ChatHistoryAction
