+++ src/gameclient/ui/store/actiontypes.ts
@@ -1,0 +1,1 @@
+export const APPEND_CHAT_HISTORY = "APPEND_CHAT_HISTORY";
